% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzles.R
\name{get_puzzle}
\alias{get_puzzle}
\title{Load a predefined puzzle as a slitherlink_grid}
\usage{
get_puzzle(name)
}
\arguments{
\item{name}{A string matching one of the names returned by list_puzzles().}
}
\value{
An object of class 'slitherlink_grid' ready to play.
}
\description{
Load a predefined puzzle as a slitherlink_grid
}
