% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzles.R
\name{list_puzzles}
\alias{list_puzzles}
\title{List available predefined puzzles}
\usage{
list_puzzles()
}
\value{
A data frame with puzzle names, difficulty, and size.
}
\description{
List available predefined puzzles
}
