% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{print.slitherlink_grid}
\alias{print.slitherlink_grid}
\title{Print a Slitherlink grid}
\usage{
\method{print}{slitherlink_grid}(x, ...)
}
\arguments{
\item{x}{An object of class 'slitherlink_grid'.}

\item{...}{Additional arguments (unused).}
}
\description{
Custom print method to display the grid in the R console.
}
\keyword{internal}
