% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{propagate}
\alias{propagate}
\title{Propagate constraints until no more deductions can be made}
\usage{
propagate(grid)
}
\arguments{
\item{grid}{An object of class 'slitherlink_grid'.}
}
\value{
A list with two elements:
\item{grid}{The updated grid after propagation.}
\item{contradiction}{TRUE if a contradiction was found.}
}
\description{
Repeatedly applies cell and node rules until the grid stabilises
(fixpoint) or a contradiction is detected.
}
