% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Launch the Slitherlink Shiny application}
\usage{
run_app()
}
\value{
Does not return; starts the Shiny app in the default browser.
}
\description{
Launch the Slitherlink Shiny application
}
